signals/requirements.txt
polars>=0.20.0
pandas>=2.0.0
requests>=2.31.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pyarrow>=14.0.0
plotly>=5.18.0
jupyter>=1.0.0
feedparser
beautifulsoup4
google-genai
python-dotenv
neo4j
rapidfuzz
torch
pytorch-lightning
pytorch-forecasting
tensorboard